+++ client/flutter/divkit/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.6-pre
 * Correct interpretation of dict in variables context
 * Add support of longtap actions
+* Remove dispose from public div-context
 
 ## 0.1.5
 * Templater support: templates can be inherited, composited, links and transitive links to fields.